+++ customWindow.py
@@ -5,35 +5,38 @@ from PyQt6.QtGui import QMovie, QPalette, QBrush, QPixmap, QDesktopServices, QIc
 from PyQt6.QtCore import Qt, QTimer, pyqtSignal
 from PyQt6 import QtCore, QtGui, QtWidgets
 
-import UI
+import skeletonWindow
 import xrayWindow
 import space_fillingWindow
 import ribbonWindow
 import amino_acidsWindow
-import customWindow
 
 from variables import decorative_blocks
 import pandas as pd
 
 import pdb_manipulation as pdbm
 import minecraft_functions as mcf
-import skeleton
-from utilUI import MyComboBox, NothingSelected, IncludedPDBPopup, MinecraftPopup, FileExplorerPopup
+import custom
 
-class SkeletonWindow(QMainWindow):
+#from skeletonWindow import SkeletonWindow
+from utilUI import IncludedPDBPopup, NothingSelected, MyComboBox, MinecraftPopup, FileExplorerPopup
+
+
+class CustomWindow(QMainWindow):
     def __init__(self):
         super().__init__()
         self.user_pdb_file = None
         self.user_minecraft_save = None
-        self.setWindowTitle("Skeleton Mode")
+        self.setWindowTitle("Custom mode")
+        self.setFixedSize(607, 411)
         self.setWindowIcon(QIcon('images/icons/logo.png'))
 
-        self.resize(450, 411)
         # Set style to Fusion
         #self.setStyle("Fusion")
         self.centralwidget = QtWidgets.QWidget(parent=self)
         self.centralwidget.setObjectName("centralwidget")
         self.switchModeLabel = QtWidgets.QLabel(parent=self.centralwidget)
+        self.switchModeLabel.setGeometry(QtCore.QRect(0, 0, 101, 31))
         font = QtGui.QFont()
         font.setPointSize(10)
         font.setBold(True)
@@ -44,67 +47,67 @@ class SkeletonWindow(QMainWindow):
         self.switchModeLabel.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
         self.switchModeLabel.setObjectName("switchModeLabel")
         self.vSepLine = QtWidgets.QFrame(parent=self.centralwidget)
-
+        self.vSepLine.setGeometry(QtCore.QRect(90, 0, 20, 431))
         self.vSepLine.setFrameShadow(QtWidgets.QFrame.Shadow.Sunken)
         self.vSepLine.setLineWidth(2)
         self.vSepLine.setFrameShape(QtWidgets.QFrame.Shape.VLine)
         self.vSepLine.setObjectName("vSepLine")
         self.CustomMode = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.CustomMode.setGeometry(QtCore.QRect(10, 30, 75, 23))
         font = QtGui.QFont()
         font.setPointSize(11)
         self.CustomMode.setFont(font)
         self.CustomMode.setObjectName("CustomMode")
         self.SkeletonMode = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.SkeletonMode.setGeometry(QtCore.QRect(10, 60, 75, 23))
         font = QtGui.QFont()
         font.setPointSize(11)
         self.SkeletonMode.setFont(font)
         self.SkeletonMode.setObjectName("SkeletonMode")
         self.XRayMode = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.XRayMode.setGeometry(QtCore.QRect(10, 90, 75, 23))
         font = QtGui.QFont()
         font.setPointSize(11)
         self.XRayMode.setFont(font)
         self.XRayMode.setObjectName("XRayMode")
         self.SpaceFillingMode = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.SpaceFillingMode.setGeometry(QtCore.QRect(10, 120, 75, 23))
         font = QtGui.QFont()
         font.setPointSize(8)
         self.SpaceFillingMode.setFont(font)
         self.SpaceFillingMode.setObjectName("SpaceFillingMode")
         self.AminoAcidMode = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.AminoAcidMode.setGeometry(QtCore.QRect(10, 180, 75, 23))
         font = QtGui.QFont()
         font.setPointSize(8)
         self.AminoAcidMode.setFont(font)
         self.AminoAcidMode.setObjectName("AminoAcidMode")
         self.RibbonMode = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.RibbonMode.setGeometry(QtCore.QRect(10, 150, 75, 23))
         font = QtGui.QFont()
         font.setPointSize(11)
         self.RibbonMode.setFont(font)
         self.RibbonMode.setObjectName("RibbonMode")
         self.github = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.github.setGeometry(QtCore.QRect(10, 290, 75, 31))
         font = QtGui.QFont()
         font.setPointSize(10)
         self.github.setFont(font)
         self.github.setObjectName("github")
         self.help = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.help.setGeometry(QtCore.QRect(10, 250, 75, 31))
         font = QtGui.QFont()
         font.setPointSize(10)
         self.help.setFont(font)
         self.help.setObjectName("help")
         self.rcsbButton = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.rcsbButton.setGeometry(QtCore.QRect(10, 360, 75, 31))
         font = QtGui.QFont()
         font.setPointSize(9)
         self.rcsbButton.setFont(font)
         self.rcsbButton.setObjectName("rcsbButton")
         self.mc2pdbLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.mc2pdbLabel.setGeometry(QtCore.QRect(10, 220, 71, 21))
         font = QtGui.QFont()
         font.setPointSize(10)
         font.setBold(False)
@@ -115,7 +118,7 @@ class SkeletonWindow(QMainWindow):
         self.mc2pdbLabel.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
         self.mc2pdbLabel.setObjectName("mc2pdbLabel")
         self.pdbDatabaseLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.pdbDatabaseLabel.setGeometry(QtCore.QRect(0, 330, 91, 21))
         font = QtGui.QFont()
         font.setPointSize(10)
         font.setBold(False)
@@ -126,17 +129,17 @@ class SkeletonWindow(QMainWindow):
         self.pdbDatabaseLabel.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
         self.pdbDatabaseLabel.setObjectName("pdbDatabaseLabel")
         self.modeInfoHLine = QtWidgets.QFrame(parent=self.centralwidget)
-
+        self.modeInfoHLine.setGeometry(QtCore.QRect(10, 210, 71, 16))
         self.modeInfoHLine.setFrameShape(QtWidgets.QFrame.Shape.HLine)
         self.modeInfoHLine.setFrameShadow(QtWidgets.QFrame.Shadow.Sunken)
         self.modeInfoHLine.setObjectName("modeInfoHLine")
         self.infoDatabaseHLine = QtWidgets.QFrame(parent=self.centralwidget)
-
+        self.infoDatabaseHLine.setGeometry(QtCore.QRect(10, 320, 71, 16))
         self.infoDatabaseHLine.setFrameShape(QtWidgets.QFrame.Shape.HLine)
         self.infoDatabaseHLine.setFrameShadow(QtWidgets.QFrame.Shadow.Sunken)
         self.infoDatabaseHLine.setObjectName("infoDatabaseHLine")
         self.bg = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.bg.setGeometry(QtCore.QRect(-90, -50, 781, 461))
         font = QtGui.QFont()
         font.setPointSize(7)
         self.bg.setFont(font)
@@ -336,19 +339,303 @@ class SkeletonWindow(QMainWindow):
         icon64 = QtGui.QIcon()
         icon64.addPixmap(QtGui.QPixmap("images/icons/white_wool.png"), QtGui.QIcon.Mode.Normal, QtGui.QIcon.State.Off)
 
+        self.cColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
+        self.cColorLabel.setGeometry(QtCore.QRect(110, 10, 106, 21))
+        self.cColorLabel.setObjectName("cColorLabel")
+        self.oColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
+        self.oColorLabel.setGeometry(QtCore.QRect(110, 40, 106, 21))
+        self.oColorLabel.setObjectName("oColorLabel")
+        self.nColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
+        self.nColorLabel.setGeometry(QtCore.QRect(110, 70, 106, 21))
+        self.nColorLabel.setObjectName("nColorLabel")
+        self.sColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
+        self.sColorLabel.setGeometry(QtCore.QRect(110, 100, 106, 21))
+        self.sColorLabel.setObjectName("sColorLabel")
+        self.pColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
+        self.pColorLabel.setGeometry(QtCore.QRect(110, 130, 110, 21))
+        self.pColorLabel.setObjectName("pColorLabel")
         self.backboneColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
-
+        self.backboneColorLabel.setGeometry(QtCore.QRect(110, 240, 120, 21))
         self.backboneColorLabel.setObjectName("backboneColorLabel")
         self.sidechainColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.sidechainColorLabel.setGeometry(QtCore.QRect(110, 210, 120, 21))
         self.sidechainColorLabel.setObjectName("sidechainColorLabel")
         self.otherColorLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.otherColorLabel.setGeometry(QtCore.QRect(110, 160, 121, 21))
         self.otherColorLabel.setObjectName("otherColorLabel")
+        #self.oColorBox = QtWidgets.QComboBox(parent=self.centralwidget)
+        self.oColorBox = MyComboBox(self.centralwidget)
+        self.oColorBox.setGeometry(QtCore.QRect(240, 40, 175, 22))
+        self.oColorBox.setObjectName("oColorBox")
+        self.oColorBox.setEditable(True)
+
+        self.oColorBox.addItem(icon2, "red_concrete")
+        self.oColorBox.addItem(icon3, "orange_concrete")
+        self.oColorBox.addItem(icon4, "yellow_concrete")
+        self.oColorBox.addItem(icon5, "lime_concrete")
+        self.oColorBox.addItem(icon6, "green_concrete")
+        self.oColorBox.addItem(icon7, "cyan_concrete")
+        self.oColorBox.addItem(icon8, "light_blue_concrete")
+        self.oColorBox.addItem(icon9, "blue_concrete")
+        self.oColorBox.addItem(icon10, "purple_concrete")
+        self.oColorBox.addItem(icon11, "magenta_concrete")
+        self.oColorBox.addItem(icon12, "pink_concrete")
+        self.oColorBox.addItem(icon13, "brown_concrete")
+        self.oColorBox.addItem(icon1, "black_concrete")
+        self.oColorBox.addItem(icon14, "gray_concrete")
+        self.oColorBox.addItem(icon15, "light_gray_concrete")
+        self.oColorBox.addItem(icon16, "white_concrete")
+        self.oColorBox.addItem(icon17, "red_glazed_terracotta")
+        self.oColorBox.addItem(icon18, "orange_glazed_terracotta")
+        self.oColorBox.addItem(icon19, "yellow_glazed_terracotta")
+        self.oColorBox.addItem(icon20, "lime_glazed_terracotta")
+        self.oColorBox.addItem(icon21, "green_glazed_terracotta")
+        self.oColorBox.addItem(icon22, "cyan_glazed_terracotta")
+        self.oColorBox.addItem(icon23, "light_blue_glazed_terracotta")
+        self.oColorBox.addItem(icon24, "blue_glazed_terracotta")
+        self.oColorBox.addItem(icon25, "purple_glazed_terracotta")
+        self.oColorBox.addItem(icon26, "magenta_glazed_terracotta")
+        self.oColorBox.addItem(icon27, "pink_glazed_terracotta")
+        self.oColorBox.addItem(icon28, "brown_glazed_terracotta")
+        self.oColorBox.addItem(icon29, "black_glazed_terracotta")
+        self.oColorBox.addItem(icon30, "gray_glazed_terracotta")
+        self.oColorBox.addItem(icon31, "light_gray_glazed_terracotta")
+        self.oColorBox.addItem(icon32, "white_glazed_terracotta")
+        self.oColorBox.addItem(icon33, "red_terracotta")
+        self.oColorBox.addItem(icon34, "orange_terracotta")
+        self.oColorBox.addItem(icon35, "yellow_terracotta")
+        self.oColorBox.addItem(icon36, "lime_terracotta")
+        self.oColorBox.addItem(icon37, "green_terracotta")
+        self.oColorBox.addItem(icon38, "cyan_terracotta")
+        self.oColorBox.addItem(icon39, "light_blue_terracotta")
+        self.oColorBox.addItem(icon40, "blue_terracotta")
+        self.oColorBox.addItem(icon41, "purple_terracotta")
+        self.oColorBox.addItem(icon42, "magenta_terracotta")
+        self.oColorBox.addItem(icon43, "pink_terracotta")
+        self.oColorBox.addItem(icon44, "brown_terracotta")
+        self.oColorBox.addItem(icon45, "black_terracotta")
+        self.oColorBox.addItem(icon46, "gray_terracotta")
+        self.oColorBox.addItem(icon47, "light_gray_terracotta")
+        self.oColorBox.addItem(icon48, "white_terracotta")
+        self.oColorBox.addItem(icon49, "red_wool")
+        self.oColorBox.addItem(icon50, "orange_wool")
+        self.oColorBox.addItem(icon51, "yellow_wool")
+        self.oColorBox.addItem(icon52, "lime_wool")
+        self.oColorBox.addItem(icon53, "green_wool")
+        self.oColorBox.addItem(icon54, "cyan_wool")
+        self.oColorBox.addItem(icon55, "light_blue_wool")
+        self.oColorBox.addItem(icon56, "blue_wool")
+        self.oColorBox.addItem(icon57, "purple_wool")
+        self.oColorBox.addItem(icon58, "magenta_wool")
+        self.oColorBox.addItem(icon59, "pink_wool")
+        self.oColorBox.addItem(icon60, "brown_wool")
+        self.oColorBox.addItem(icon61, "black_wool")
+        self.oColorBox.addItem(icon62, "gray_wool")
+        self.oColorBox.addItem(icon63, "light_gray_wool")
+        self.oColorBox.addItem(icon64, "wool")
+        self.oColorBox.addItem(icon2, "red_stained_glass")
+        self.oColorBox.addItem(icon3, "orange_stained_glass")
+        self.oColorBox.addItem(icon4, "yellow_stained_glass")
+        self.oColorBox.addItem(icon5, "lime_stained_glass")
+        self.oColorBox.addItem(icon6, "green_stained_glass")
+        self.oColorBox.addItem(icon7, "cyan_stained_glass")
+        self.oColorBox.addItem(icon8, "light_blue_stained_glass")
+        self.oColorBox.addItem(icon9, "blue_stained_glass")
+        self.oColorBox.addItem(icon10, "purple_stained_glass")
+        self.oColorBox.addItem(icon11, "magenta_stained_glass")
+        self.oColorBox.addItem(icon12, "pink_stained_glass")
+        self.oColorBox.addItem(icon13, "brown_stained_glass")
+        self.oColorBox.addItem(icon1, "black_stained_glass")
+        self.oColorBox.addItem(icon14, "gray_stained_glass")
+        self.oColorBox.addItem(icon15, "light_stained_glass")
+        self.oColorBox.addItem(icon16, "white_stained_glass")
+        self.oColorBox.insertSeparator(16)
+        self.oColorBox.insertSeparator(33)
+        self.oColorBox.insertSeparator(50)
+        self.oColorBox.insertSeparator(67)
+
+        self.nColorBox = MyComboBox(self.centralwidget)
+        self.nColorBox.setGeometry(QtCore.QRect(240, 70, 175, 22))
+        self.nColorBox.setObjectName("nColorBox")
+        self.nColorBox.setEditable(True)
+
+        self.nColorBox.addItem(icon9, "blue_concrete")
+        self.nColorBox.addItem(icon2, "red_concrete")
+        self.nColorBox.addItem(icon3, "orange_concrete")
+        self.nColorBox.addItem(icon4, "yellow_concrete")
+        self.nColorBox.addItem(icon5, "lime_concrete")
+        self.nColorBox.addItem(icon6, "green_concrete")
+        self.nColorBox.addItem(icon7, "cyan_concrete")
+        self.nColorBox.addItem(icon8, "light_blue_concrete")
+        self.nColorBox.addItem(icon10, "purple_concrete")
+        self.nColorBox.addItem(icon11, "magenta_concrete")
+        self.nColorBox.addItem(icon12, "pink_concrete")
+        self.nColorBox.addItem(icon13, "brown_concrete")
+        self.nColorBox.addItem(icon1, "black_concrete")
+        self.nColorBox.addItem(icon14, "gray_concrete")
+        self.nColorBox.addItem(icon15, "light_gray_concrete")
+        self.nColorBox.addItem(icon16, "white_concrete")
+        self.nColorBox.addItem(icon17, "red_glazed_terracotta")
+        self.nColorBox.addItem(icon18, "orange_glazed_terracotta")
+        self.nColorBox.addItem(icon19, "yellow_glazed_terracotta")
+        self.nColorBox.addItem(icon20, "lime_glazed_terracotta")
+        self.nColorBox.addItem(icon21, "green_glazed_terracotta")
+        self.nColorBox.addItem(icon22, "cyan_glazed_terracotta")
+        self.nColorBox.addItem(icon23, "light_blue_glazed_terracotta")
+        self.nColorBox.addItem(icon24, "blue_glazed_terracotta")
+        self.nColorBox.addItem(icon25, "purple_glazed_terracotta")
+        self.nColorBox.addItem(icon26, "magenta_glazed_terracotta")
+        self.nColorBox.addItem(icon27, "pink_glazed_terracotta")
+        self.nColorBox.addItem(icon28, "brown_glazed_terracotta")
+        self.nColorBox.addItem(icon29, "black_glazed_terracotta")
+        self.nColorBox.addItem(icon30, "gray_glazed_terracotta")
+        self.nColorBox.addItem(icon31, "light_gray_glazed_terracotta")
+        self.nColorBox.addItem(icon32, "white_glazed_terracotta")
+        self.nColorBox.addItem(icon33, "red_terracotta")
+        self.nColorBox.addItem(icon34, "orange_terracotta")
+        self.nColorBox.addItem(icon35, "yellow_terracotta")
+        self.nColorBox.addItem(icon36, "lime_terracotta")
+        self.nColorBox.addItem(icon37, "green_terracotta")
+        self.nColorBox.addItem(icon38, "cyan_terracotta")
+        self.nColorBox.addItem(icon39, "light_blue_terracotta")
+        self.nColorBox.addItem(icon40, "blue_terracotta")
+        self.nColorBox.addItem(icon41, "purple_terracotta")
+        self.nColorBox.addItem(icon42, "magenta_terracotta")
+        self.nColorBox.addItem(icon43, "pink_terracotta")
+        self.nColorBox.addItem(icon44, "brown_terracotta")
+        self.nColorBox.addItem(icon45, "black_terracotta")
+        self.nColorBox.addItem(icon46, "gray_terracotta")
+        self.nColorBox.addItem(icon47, "light_gray_terracotta")
+        self.nColorBox.addItem(icon48, "white_terracotta")
+        self.nColorBox.addItem(icon49, "red_wool")
+        self.nColorBox.addItem(icon50, "orange_wool")
+        self.nColorBox.addItem(icon51, "yellow_wool")
+        self.nColorBox.addItem(icon52, "lime_wool")
+        self.nColorBox.addItem(icon53, "green_wool")
+        self.nColorBox.addItem(icon54, "cyan_wool")
+        self.nColorBox.addItem(icon55, "light_blue_wool")
+        self.nColorBox.addItem(icon56, "blue_wool")
+        self.nColorBox.addItem(icon57, "purple_wool")
+        self.nColorBox.addItem(icon58, "magenta_wool")
+        self.nColorBox.addItem(icon59, "pink_wool")
+        self.nColorBox.addItem(icon60, "brown_wool")
+        self.nColorBox.addItem(icon61, "black_wool")
+        self.nColorBox.addItem(icon62, "gray_wool")
+        self.nColorBox.addItem(icon63, "light_gray_wool")
+        self.nColorBox.addItem(icon64, "wool")
+        self.nColorBox.addItem(icon2, "red_stained_glass")
+        self.nColorBox.addItem(icon3, "orange_stained_glass")
+        self.nColorBox.addItem(icon4, "yellow_stained_glass")
+        self.nColorBox.addItem(icon5, "lime_stained_glass")
+        self.nColorBox.addItem(icon6, "green_stained_glass")
+        self.nColorBox.addItem(icon7, "cyan_stained_glass")
+        self.nColorBox.addItem(icon8, "light_blue_stained_glass")
+        self.nColorBox.addItem(icon9, "blue_stained_glass")
+        self.nColorBox.addItem(icon10, "purple_stained_glass")
+        self.nColorBox.addItem(icon11, "magenta_stained_glass")
+        self.nColorBox.addItem(icon12, "pink_stained_glass")
+        self.nColorBox.addItem(icon13, "brown_stained_glass")
+        self.nColorBox.addItem(icon1, "black_stained_glass")
+        self.nColorBox.addItem(icon14, "gray_stained_glass")
+        self.nColorBox.addItem(icon15, "light_stained_glass")
+        self.nColorBox.addItem(icon16, "white_stained_glass")
+        self.nColorBox.insertSeparator(16)
+        self.nColorBox.insertSeparator(33)
+        self.nColorBox.insertSeparator(50)
+        self.nColorBox.insertSeparator(67)
+
+        self.pColorBox = MyComboBox(self.centralwidget)
+        self.pColorBox.setGeometry(QtCore.QRect(240, 130, 175, 22))
+        self.pColorBox.setObjectName("pColorBox")
+        self.pColorBox.setEditable(True)
+
+        self.pColorBox.addItem(icon6, "green_concrete")
+        self.pColorBox.addItem(icon2, "red_concrete")
+        self.pColorBox.addItem(icon3, "orange_concrete")
+        self.pColorBox.addItem(icon4, "yellow_concrete")
+        self.pColorBox.addItem(icon5, "lime_concrete")
+        self.pColorBox.addItem(icon7, "cyan_concrete")
+        self.pColorBox.addItem(icon8, "light_blue_concrete")
+        self.pColorBox.addItem(icon9, "blue_concrete")
+        self.pColorBox.addItem(icon10, "purple_concrete")
+        self.pColorBox.addItem(icon11, "magenta_concrete")
+        self.pColorBox.addItem(icon12, "pink_concrete")
+        self.pColorBox.addItem(icon13, "brown_concrete")
+        self.pColorBox.addItem(icon1, "black_concrete")
+        self.pColorBox.addItem(icon14, "gray_concrete")
+        self.pColorBox.addItem(icon15, "light_gray_concrete")
+        self.pColorBox.addItem(icon16, "white_concrete")
+        self.pColorBox.addItem(icon17, "red_glazed_terracotta")
+        self.pColorBox.addItem(icon18, "orange_glazed_terracotta")
+        self.pColorBox.addItem(icon19, "yellow_glazed_terracotta")
+        self.pColorBox.addItem(icon20, "lime_glazed_terracotta")
+        self.pColorBox.addItem(icon21, "green_glazed_terracotta")
+        self.pColorBox.addItem(icon22, "cyan_glazed_terracotta")
+        self.pColorBox.addItem(icon23, "light_blue_glazed_terracotta")
+        self.pColorBox.addItem(icon24, "blue_glazed_terracotta")
+        self.pColorBox.addItem(icon25, "purple_glazed_terracotta")
+        self.pColorBox.addItem(icon26, "magenta_glazed_terracotta")
+        self.pColorBox.addItem(icon27, "pink_glazed_terracotta")
+        self.pColorBox.addItem(icon28, "brown_glazed_terracotta")
+        self.pColorBox.addItem(icon29, "black_glazed_terracotta")
+        self.pColorBox.addItem(icon30, "gray_glazed_terracotta")
+        self.pColorBox.addItem(icon31, "light_gray_glazed_terracotta")
+        self.pColorBox.addItem(icon32, "white_glazed_terracotta")
+        self.pColorBox.addItem(icon33, "red_terracotta")
+        self.pColorBox.addItem(icon34, "orange_terracotta")
+        self.pColorBox.addItem(icon35, "yellow_terracotta")
+        self.pColorBox.addItem(icon36, "lime_terracotta")
+        self.pColorBox.addItem(icon37, "green_terracotta")
+        self.pColorBox.addItem(icon38, "cyan_terracotta")
+        self.pColorBox.addItem(icon39, "light_blue_terracotta")
+        self.pColorBox.addItem(icon40, "blue_terracotta")
+        self.pColorBox.addItem(icon41, "purple_terracotta")
+        self.pColorBox.addItem(icon42, "magenta_terracotta")
+        self.pColorBox.addItem(icon43, "pink_terracotta")
+        self.pColorBox.addItem(icon44, "brown_terracotta")
+        self.pColorBox.addItem(icon45, "black_terracotta")
+        self.pColorBox.addItem(icon46, "gray_terracotta")
+        self.pColorBox.addItem(icon47, "light_gray_terracotta")
+        self.pColorBox.addItem(icon48, "white_terracotta")
+        self.pColorBox.addItem(icon49, "red_wool")
+        self.pColorBox.addItem(icon50, "orange_wool")
+        self.pColorBox.addItem(icon51, "yellow_wool")
+        self.pColorBox.addItem(icon52, "lime_wool")
+        self.pColorBox.addItem(icon53, "green_wool")
+        self.pColorBox.addItem(icon54, "cyan_wool")
+        self.pColorBox.addItem(icon55, "light_blue_wool")
+        self.pColorBox.addItem(icon56, "blue_wool")
+        self.pColorBox.addItem(icon57, "purple_wool")
+        self.pColorBox.addItem(icon58, "magenta_wool")
+        self.pColorBox.addItem(icon59, "pink_wool")
+        self.pColorBox.addItem(icon60, "brown_wool")
+        self.pColorBox.addItem(icon61, "black_wool")
+        self.pColorBox.addItem(icon62, "gray_wool")
+        self.pColorBox.addItem(icon63, "light_gray_wool")
+        self.pColorBox.addItem(icon64, "wool")
+        self.pColorBox.addItem(icon2, "red_stained_glass")
+        self.pColorBox.addItem(icon3, "orange_stained_glass")
+        self.pColorBox.addItem(icon4, "yellow_stained_glass")
+        self.pColorBox.addItem(icon5, "lime_stained_glass")
+        self.pColorBox.addItem(icon6, "green_stained_glass")
+        self.pColorBox.addItem(icon7, "cyan_stained_glass")
+        self.pColorBox.addItem(icon8, "light_blue_stained_glass")
+        self.pColorBox.addItem(icon9, "blue_stained_glass")
+        self.pColorBox.addItem(icon10, "purple_stained_glass")
+        self.pColorBox.addItem(icon11, "magenta_stained_glass")
+        self.pColorBox.addItem(icon12, "pink_stained_glass")
+        self.pColorBox.addItem(icon13, "brown_stained_glass")
+        self.pColorBox.addItem(icon1, "black_stained_glass")
+        self.pColorBox.addItem(icon14, "gray_stained_glass")
+        self.pColorBox.addItem(icon15, "light_stained_glass")
+        self.pColorBox.addItem(icon16, "white_stained_glass")
+        self.pColorBox.insertSeparator(16)
+        self.pColorBox.insertSeparator(33)
+        self.pColorBox.insertSeparator(50)
+        self.pColorBox.insertSeparator(67)
 
         self.otherColorBox = MyComboBox(self.centralwidget)
-
+        self.otherColorBox.setGeometry(QtCore.QRect(240, 160, 175, 22))
         self.otherColorBox.setObjectName("otherColorBox")
         self.otherColorBox.setEditable(True)
 
@@ -437,8 +724,188 @@ class SkeletonWindow(QMainWindow):
         self.otherColorBox.insertSeparator(50)
         self.otherColorBox.insertSeparator(67)
 
-        self.backboneColorBox = MyComboBox(self.centralwidget)
+        self.sColorBox = MyComboBox(self.centralwidget)
+        self.sColorBox.setGeometry(QtCore.QRect(240, 100, 175, 22))
+        self.sColorBox.setObjectName("sColorBox")
+        self.sColorBox.setEditable(True)
 
+        self.sColorBox.addItem(icon4, "yellow_concrete")
+        self.sColorBox.addItem(icon2, "red_concrete")
+        self.sColorBox.addItem(icon3, "orange_concrete")
+        self.sColorBox.addItem(icon5, "lime_concrete")
+        self.sColorBox.addItem(icon6, "green_concrete")
+        self.sColorBox.addItem(icon7, "cyan_concrete")
+        self.sColorBox.addItem(icon8, "light_blue_concrete")
+        self.sColorBox.addItem(icon9, "blue_concrete")
+        self.sColorBox.addItem(icon10, "purple_concrete")
+        self.sColorBox.addItem(icon11, "magenta_concrete")
+        self.sColorBox.addItem(icon12, "pink_concrete")
+        self.sColorBox.addItem(icon13, "brown_concrete")
+        self.sColorBox.addItem(icon1, "black_concrete")
+        self.sColorBox.addItem(icon14, "gray_concrete")
+        self.sColorBox.addItem(icon15, "light_gray_concrete")
+        self.sColorBox.addItem(icon16, "white_concrete")
+        self.sColorBox.addItem(icon17, "red_glazed_terracotta")
+        self.sColorBox.addItem(icon18, "orange_glazed_terracotta")
+        self.sColorBox.addItem(icon19, "yellow_glazed_terracotta")
+        self.sColorBox.addItem(icon20, "lime_glazed_terracotta")
+        self.sColorBox.addItem(icon21, "green_glazed_terracotta")
+        self.sColorBox.addItem(icon22, "cyan_glazed_terracotta")
+        self.sColorBox.addItem(icon23, "light_blue_glazed_terracotta")
+        self.sColorBox.addItem(icon24, "blue_glazed_terracotta")
+        self.sColorBox.addItem(icon25, "purple_glazed_terracotta")
+        self.sColorBox.addItem(icon26, "magenta_glazed_terracotta")
+        self.sColorBox.addItem(icon27, "pink_glazed_terracotta")
+        self.sColorBox.addItem(icon28, "brown_glazed_terracotta")
+        self.sColorBox.addItem(icon29, "black_glazed_terracotta")
+        self.sColorBox.addItem(icon30, "gray_glazed_terracotta")
+        self.sColorBox.addItem(icon31, "light_gray_glazed_terracotta")
+        self.sColorBox.addItem(icon32, "white_glazed_terracotta")
+        self.sColorBox.addItem(icon33, "red_terracotta")
+        self.sColorBox.addItem(icon34, "orange_terracotta")
+        self.sColorBox.addItem(icon35, "yellow_terracotta")
+        self.sColorBox.addItem(icon36, "lime_terracotta")
+        self.sColorBox.addItem(icon37, "green_terracotta")
+        self.sColorBox.addItem(icon38, "cyan_terracotta")
+        self.sColorBox.addItem(icon39, "light_blue_terracotta")
+        self.sColorBox.addItem(icon40, "blue_terracotta")
+        self.sColorBox.addItem(icon41, "purple_terracotta")
+        self.sColorBox.addItem(icon42, "magenta_terracotta")
+        self.sColorBox.addItem(icon43, "pink_terracotta")
+        self.sColorBox.addItem(icon44, "brown_terracotta")
+        self.sColorBox.addItem(icon45, "black_terracotta")
+        self.sColorBox.addItem(icon46, "gray_terracotta")
+        self.sColorBox.addItem(icon47, "light_gray_terracotta")
+        self.sColorBox.addItem(icon48, "white_terracotta")
+        self.sColorBox.addItem(icon49, "red_wool")
+        self.sColorBox.addItem(icon50, "orange_wool")
+        self.sColorBox.addItem(icon51, "yellow_wool")
+        self.sColorBox.addItem(icon52, "lime_wool")
+        self.sColorBox.addItem(icon53, "green_wool")
+        self.sColorBox.addItem(icon54, "cyan_wool")
+        self.sColorBox.addItem(icon55, "light_blue_wool")
+        self.sColorBox.addItem(icon56, "blue_wool")
+        self.sColorBox.addItem(icon57, "purple_wool")
+        self.sColorBox.addItem(icon58, "magenta_wool")
+        self.sColorBox.addItem(icon59, "pink_wool")
+        self.sColorBox.addItem(icon60, "brown_wool")
+        self.sColorBox.addItem(icon61, "black_wool")
+        self.sColorBox.addItem(icon62, "gray_wool")
+        self.sColorBox.addItem(icon63, "light_gray_wool")
+        self.sColorBox.addItem(icon64, "wool")
+        self.sColorBox.addItem(icon2, "red_stained_glass")
+        self.sColorBox.addItem(icon3, "orange_stained_glass")
+        self.sColorBox.addItem(icon4, "yellow_stained_glass")
+        self.sColorBox.addItem(icon5, "lime_stained_glass")
+        self.sColorBox.addItem(icon6, "green_stained_glass")
+        self.sColorBox.addItem(icon7, "cyan_stained_glass")
+        self.sColorBox.addItem(icon8, "light_blue_stained_glass")
+        self.sColorBox.addItem(icon9, "blue_stained_glass")
+        self.sColorBox.addItem(icon10, "purple_stained_glass")
+        self.sColorBox.addItem(icon11, "magenta_stained_glass")
+        self.sColorBox.addItem(icon12, "pink_stained_glass")
+        self.sColorBox.addItem(icon13, "brown_stained_glass")
+        self.sColorBox.addItem(icon1, "black_stained_glass")
+        self.sColorBox.addItem(icon14, "gray_stained_glass")
+        self.sColorBox.addItem(icon15, "light_stained_glass")
+        self.sColorBox.addItem(icon16, "white_stained_glass")
+        self.sColorBox.insertSeparator(16)
+        self.sColorBox.insertSeparator(33)
+        self.sColorBox.insertSeparator(50)
+        self.sColorBox.insertSeparator(67)
+
+        self.cColorBox = MyComboBox(self.centralwidget)
+        self.cColorBox.setGeometry(QtCore.QRect(240, 10, 175, 22))
+        self.cColorBox.setObjectName("cColorBox")
+        self.cColorBox.setEditable(True)
+
+        self.cColorBox.addItem(icon1, "black_concrete")
+        self.cColorBox.addItem(icon2, "red_concrete")
+        self.cColorBox.addItem(icon3, "orange_concrete")
+        self.cColorBox.addItem(icon4, "yellow_concrete")
+        self.cColorBox.addItem(icon5, "lime_concrete")
+        self.cColorBox.addItem(icon6, "green_concrete")
+        self.cColorBox.addItem(icon7, "cyan_concrete")
+        self.cColorBox.addItem(icon8, "light_blue_concrete")
+        self.cColorBox.addItem(icon9, "blue_concrete")
+        self.cColorBox.addItem(icon10, "purple_concrete")
+        self.cColorBox.addItem(icon11, "magenta_concrete")
+        self.cColorBox.addItem(icon12, "pink_concrete")
+        self.cColorBox.addItem(icon13, "brown_concrete")
+        self.cColorBox.addItem(icon14, "gray_concrete")
+        self.cColorBox.addItem(icon15, "light_gray_concrete")
+        self.cColorBox.addItem(icon16, "white_concrete")
+        self.cColorBox.addItem(icon17, "red_glazed_terracotta")
+        self.cColorBox.addItem(icon18, "orange_glazed_terracotta")
+        self.cColorBox.addItem(icon19, "yellow_glazed_terracotta")
+        self.cColorBox.addItem(icon20, "lime_glazed_terracotta")
+        self.cColorBox.addItem(icon21, "green_glazed_terracotta")
+        self.cColorBox.addItem(icon22, "cyan_glazed_terracotta")
+        self.cColorBox.addItem(icon23, "light_blue_glazed_terracotta")
+        self.cColorBox.addItem(icon24, "blue_glazed_terracotta")
+        self.cColorBox.addItem(icon25, "purple_glazed_terracotta")
+        self.cColorBox.addItem(icon26, "magenta_glazed_terracotta")
+        self.cColorBox.addItem(icon27, "pink_glazed_terracotta")
+        self.cColorBox.addItem(icon28, "brown_glazed_terracotta")
+        self.cColorBox.addItem(icon29, "black_glazed_terracotta")
+        self.cColorBox.addItem(icon30, "gray_glazed_terracotta")
+        self.cColorBox.addItem(icon31, "light_gray_glazed_terracotta")
+        self.cColorBox.addItem(icon32, "white_glazed_terracotta")
+        self.cColorBox.addItem(icon33, "red_terracotta")
+        self.cColorBox.addItem(icon34, "orange_terracotta")
+        self.cColorBox.addItem(icon35, "yellow_terracotta")
+        self.cColorBox.addItem(icon36, "lime_terracotta")
+        self.cColorBox.addItem(icon37, "green_terracotta")
+        self.cColorBox.addItem(icon38, "cyan_terracotta")
+        self.cColorBox.addItem(icon39, "light_blue_terracotta")
+        self.cColorBox.addItem(icon40, "blue_terracotta")
+        self.cColorBox.addItem(icon41, "purple_terracotta")
+        self.cColorBox.addItem(icon42, "magenta_terracotta")
+        self.cColorBox.addItem(icon43, "pink_terracotta")
+        self.cColorBox.addItem(icon44, "brown_terracotta")
+        self.cColorBox.addItem(icon45, "black_terracotta")
+        self.cColorBox.addItem(icon46, "gray_terracotta")
+        self.cColorBox.addItem(icon47, "light_gray_terracotta")
+        self.cColorBox.addItem(icon48, "white_terracotta")
+        self.cColorBox.addItem(icon49, "red_wool")
+        self.cColorBox.addItem(icon50, "orange_wool")
+        self.cColorBox.addItem(icon51, "yellow_wool")
+        self.cColorBox.addItem(icon52, "lime_wool")
+        self.cColorBox.addItem(icon53, "green_wool")
+        self.cColorBox.addItem(icon54, "cyan_wool")
+        self.cColorBox.addItem(icon55, "light_blue_wool")
+        self.cColorBox.addItem(icon56, "blue_wool")
+        self.cColorBox.addItem(icon57, "purple_wool")
+        self.cColorBox.addItem(icon58, "magenta_wool")
+        self.cColorBox.addItem(icon59, "pink_wool")
+        self.cColorBox.addItem(icon60, "brown_wool")
+        self.cColorBox.addItem(icon61, "black_wool")
+        self.cColorBox.addItem(icon62, "gray_wool")
+        self.cColorBox.addItem(icon63, "light_gray_wool")
+        self.cColorBox.addItem(icon64, "wool")
+        self.cColorBox.addItem(icon2, "red_stained_glass")
+        self.cColorBox.addItem(icon3, "orange_stained_glass")
+        self.cColorBox.addItem(icon4, "yellow_stained_glass")
+        self.cColorBox.addItem(icon5, "lime_stained_glass")
+        self.cColorBox.addItem(icon6, "green_stained_glass")
+        self.cColorBox.addItem(icon7, "cyan_stained_glass")
+        self.cColorBox.addItem(icon8, "light_blue_stained_glass")
+        self.cColorBox.addItem(icon9, "blue_stained_glass")
+        self.cColorBox.addItem(icon10, "purple_stained_glass")
+        self.cColorBox.addItem(icon11, "magenta_stained_glass")
+        self.cColorBox.addItem(icon12, "pink_stained_glass")
+        self.cColorBox.addItem(icon13, "brown_stained_glass")
+        self.cColorBox.addItem(icon1, "black_stained_glass")
+        self.cColorBox.addItem(icon14, "gray_stained_glass")
+        self.cColorBox.addItem(icon15, "light_stained_glass")
+        self.cColorBox.addItem(icon16, "white_stained_glass")
+        self.cColorBox.insertSeparator(16)
+        self.cColorBox.insertSeparator(33)
+        self.cColorBox.insertSeparator(50)
+        self.cColorBox.insertSeparator(67)
+
+        self.backboneColorBox = MyComboBox(self.centralwidget)
+        self.backboneColorBox.setGeometry(QtCore.QRect(240, 240, 175, 22))
         self.backboneColorBox.setObjectName("backboneColorBox")
         self.backboneColorBox.setEditable(True)
 
@@ -528,7 +995,7 @@ class SkeletonWindow(QMainWindow):
         self.backboneColorBox.insertSeparator(67)
 
         self.sidechainColorBox = MyComboBox(self.centralwidget)
-
+        self.sidechainColorBox.setGeometry(QtCore.QRect(240, 210, 175, 22))
         self.sidechainColorBox.setObjectName("sidechainColorBox")
         self.sidechainColorBox.setEditable(True)
 
@@ -618,24 +1085,24 @@ class SkeletonWindow(QMainWindow):
         self.sidechainColorBox.insertSeparator(67)
 
         self.aScaleLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.aScaleLabel.setGeometry(QtCore.QRect(440, 100, 61, 21))
         self.aScaleLabel.setObjectName("aScaleLabel")
-        # self.showAtomsCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-        # self.showAtomsCheck.setGeometry(QtCore.QRect(440, 10, 121, 17))
-        # self.showAtomsCheck.setChecked(True)
-        # self.showAtomsCheck.setObjectName("showAtomsCheck")
-        # self.showAtomsCheck.setToolTip("Show the atoms of the main models.")
+        self.showAtomsCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
+        self.showAtomsCheck.setGeometry(QtCore.QRect(440, 10, 121, 17))
+        self.showAtomsCheck.setChecked(True)
+        self.showAtomsCheck.setObjectName("showAtomsCheck")
+        self.showAtomsCheck.setToolTip("Show the atoms of the main models.")
         self.otherMoleculeCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-
+        self.otherMoleculeCheck.setGeometry(QtCore.QRect(440, 40, 131, 17))
         self.otherMoleculeCheck.setChecked(True)
         self.otherMoleculeCheck.setObjectName("otherMoleculeCheck")
         self.otherMoleculeCheck.setToolTip("Check to show other non-protein, DNA, or RNA molecules.")
-        #self.meshCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-        #self.meshCheck.setGeometry(QtCore.QRect(440, 70, 151, 17))
-        #self.meshCheck.setObjectName("meshCheck")
-        #self.meshCheck.setToolTip("Check to show mesh-style atoms: many fewer blocks")
+        self.meshCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
+        self.meshCheck.setGeometry(QtCore.QRect(440, 70, 151, 17))
+        self.meshCheck.setObjectName("meshCheck")
+        self.meshCheck.setToolTip("Check to show mesh-style atoms: many fewer blocks")
         self.aScaleSpinBox = QtWidgets.QDoubleSpinBox(parent=self.centralwidget)
-
+        self.aScaleSpinBox.setGeometry(QtCore.QRect(530, 100, 62, 22))
         self.aScaleSpinBox.setDecimals(1)
         self.aScaleSpinBox.setMinimum(1.0)
         self.aScaleSpinBox.setMaximum(50.0)
@@ -644,25 +1111,25 @@ class SkeletonWindow(QMainWindow):
         self.aScaleSpinBox.setObjectName("aScaleSpinBox")
         self.aScaleSpinBox.setToolTip("Change the diameter (rounded up) of each atom.")
         self.showBackboneCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-
+        self.showBackboneCheck.setGeometry(QtCore.QRect(440, 240, 121, 17))
         self.showBackboneCheck.setChecked(True)
         self.showBackboneCheck.setObjectName("showBackboneCheck")
         self.showBackboneCheck.setToolTip("Show the N-C-C backbone of the main models.")
         self.showSidechainCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-
+        self.showSidechainCheck.setGeometry(QtCore.QRect(440, 210, 121, 17))
         self.showSidechainCheck.setChecked(True)
         self.showSidechainCheck.setObjectName("showSidechainCheck")
         self.showSidechainCheck.setToolTip("Show amino acid R-groups")
         self.colorByBackboneCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-
+        self.colorByBackboneCheck.setGeometry(QtCore.QRect(240, 270, 155, 21))
         self.colorByBackboneCheck.setChecked(False)
         self.colorByBackboneCheck.setObjectName("colorByBackboneCheck")
         self.colorByBackboneCheck.setToolTip("Color the backbones of the main models by the molecule number.")
         self.pScaleLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.pScaleLabel.setGeometry(QtCore.QRect(440, 130, 71, 21))
         self.pScaleLabel.setObjectName("pScaleLabel")
         self.pScaleSpinBox = QtWidgets.QDoubleSpinBox(parent=self.centralwidget)
-
+        self.pScaleSpinBox.setGeometry(QtCore.QRect(530, 130, 62, 22))
         self.pScaleSpinBox.setDecimals(1)
         self.pScaleSpinBox.setMinimum(1.0)
         self.pScaleSpinBox.setMaximum(50.0)
@@ -670,10 +1137,10 @@ class SkeletonWindow(QMainWindow):
         self.pScaleSpinBox.setObjectName("pScaleSpinBox")
         self.pScaleSpinBox.setToolTip("Scale the entire model by this factor.")
         self.bScaleLabel = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.bScaleLabel.setGeometry(QtCore.QRect(440, 270, 111, 21))
         self.bScaleLabel.setObjectName("bScaleLabel")
         self.backboneScaleSpinBox = QtWidgets.QDoubleSpinBox(parent=self.centralwidget)
-
+        self.backboneScaleSpinBox.setGeometry(QtCore.QRect(530, 270, 62, 22))
         self.backboneScaleSpinBox.setDecimals(1)
         self.backboneScaleSpinBox.setMinimum(1.0)
         self.backboneScaleSpinBox.setMaximum(50.0)
@@ -681,25 +1148,25 @@ class SkeletonWindow(QMainWindow):
         self.backboneScaleSpinBox.setObjectName("backboneScaleSpinBox")
         self.backboneScaleSpinBox.setToolTip("Scale the width of the backbone by this factor.")
         self.selectIncludedPDBButton = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.selectIncludedPDBButton.setGeometry(QtCore.QRect(250, 320, 141, 23))
         self.selectIncludedPDBButton.setObjectName("selectIncludedPDBButton")
         self.selectMinecraftSaveButton = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.selectMinecraftSaveButton.setGeometry(QtCore.QRect(440, 320, 131, 23))
         self.selectMinecraftSaveButton.setObjectName("selectMinecraftSaveButton")
         self.simpleOutputCheck = QtWidgets.QCheckBox(parent=self.centralwidget)
-
+        self.simpleOutputCheck.setGeometry(QtCore.QRect(110, 370, 100, 31))
         self.simpleOutputCheck.setChecked(True)
         self.simpleOutputCheck.setObjectName("simpleOutputCheck")
         self.simpleOutputCheck.setToolTip("Un-select to create individual commands for each molecule")
         self.selectPDBFileButton = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.selectPDBFileButton.setGeometry(QtCore.QRect(110, 320, 91, 23))
         font = QtGui.QFont()
         font.setBold(False)
         font.setWeight(50)
         self.selectPDBFileButton.setFont(font)
         self.selectPDBFileButton.setObjectName("pushButton_10")
         self.createFunctionsButton = QtWidgets.QPushButton(parent=self.centralwidget)
-
+        self.createFunctionsButton.setGeometry(QtCore.QRect(230, 370, 181, 31))
         self.createFunctionsButton.setToolTip("Add the selected PDB file to the Minecraft world")
         font = QtGui.QFont()
         font.setPointSize(11)
@@ -708,69 +1175,19 @@ class SkeletonWindow(QMainWindow):
         self.createFunctionsButton.setFont(font)
         self.createFunctionsButton.setObjectName("pushButton_13")
         self.orText = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.orText.setGeometry(QtCore.QRect(210, 320, 21, 21))
         font = QtGui.QFont()
         font.setPointSize(9)
         self.orText.setFont(font)
         self.orText.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
         self.orText.setObjectName("or")
         self.andText = QtWidgets.QLabel(parent=self.centralwidget)
-
+        self.andText.setGeometry(QtCore.QRect(400, 320, 31, 21))
         font = QtGui.QFont()
         font.setPointSize(9)
         self.andText.setFont(font)
         self.andText.setAlignment(QtCore.Qt.AlignmentFlag.AlignCenter)
         self.andText.setObjectName("and")
-
-        self.bg.setGeometry(QtCore.QRect(-90, -50, 781, 461))
-
-        self.switchModeLabel.setGeometry(QtCore.QRect(0, 0, 101, 31))
-        self.CustomMode.setGeometry(QtCore.QRect(10, 30, 75, 23))
-        self.SkeletonMode.setGeometry(QtCore.QRect(10, 60, 75, 23))
-        self.XRayMode.setGeometry(QtCore.QRect(10, 90, 75, 23))
-        self.SpaceFillingMode.setGeometry(QtCore.QRect(10, 120, 75, 23))
-        self.RibbonMode.setGeometry(QtCore.QRect(10, 150, 75, 23))
-        self.AminoAcidMode.setGeometry(QtCore.QRect(10, 180, 75, 23))
-        self.modeInfoHLine.setGeometry(QtCore.QRect(10, 210, 71, 16))
-
-        self.mc2pdbLabel.setGeometry(QtCore.QRect(10, 220, 71, 21))
-        self.help.setGeometry(QtCore.QRect(10, 250, 75, 31))
-        self.github.setGeometry(QtCore.QRect(10, 290, 75, 31))
-        self.infoDatabaseHLine.setGeometry(QtCore.QRect(10, 320, 71, 16))
-        self.pdbDatabaseLabel.setGeometry(QtCore.QRect(0, 330, 91, 21))
-        self.rcsbButton.setGeometry(QtCore.QRect(10, 360, 75, 31))
-
-        self.vSepLine.setGeometry(QtCore.QRect(90, 0, 20, 431))
-
-        self.backboneColorLabel.setGeometry(QtCore.QRect(110, 10, 120, 21))
-        self.backboneColorBox.setGeometry(QtCore.QRect(240, 10, 175, 22))
-        self.showBackboneCheck.setGeometry(QtCore.QRect(240, 37, 121, 17))
-        self.colorByBackboneCheck.setGeometry(QtCore.QRect(240, 53, 155, 21))
-        self.bScaleLabel.setGeometry(QtCore.QRect(110, 75, 111, 21))
-        self.backboneScaleSpinBox.setGeometry(QtCore.QRect(240, 75, 62, 22))
-        self.pScaleLabel.setGeometry(QtCore.QRect(110, 100, 71, 21))
-        self.pScaleSpinBox.setGeometry(QtCore.QRect(240, 100, 62, 22))
-
-        self.sidechainColorLabel.setGeometry(QtCore.QRect(110, 140, 120, 21))
-        self.sidechainColorBox.setGeometry(QtCore.QRect(240, 140, 175, 22))
-        self.showSidechainCheck.setGeometry(QtCore.QRect(240, 165, 121, 17))
-
-        self.otherColorLabel.setGeometry(QtCore.QRect(110, 200, 121, 21))
-        self.otherColorBox.setGeometry(QtCore.QRect(240, 200, 175, 22))
-        self.otherMoleculeCheck.setGeometry(QtCore.QRect(240, 228, 131, 17))
-        self.aScaleLabel.setGeometry(QtCore.QRect(110, 250, 61, 21))
-        self.aScaleSpinBox.setGeometry(QtCore.QRect(240, 250, 62, 22))
-
-        self.selectPDBFileButton.setGeometry(QtCore.QRect(110, 290, 91, 23))
-        self.orText.setGeometry(QtCore.QRect(210, 290, 31, 21))
-        self.selectIncludedPDBButton.setGeometry(QtCore.QRect(250, 290, 141, 23))
-
-        self.andText.setGeometry(QtCore.QRect(210, 325, 31, 21))
-        self.selectMinecraftSaveButton.setGeometry(QtCore.QRect(250, 325, 141, 23))
-
-        self.simpleOutputCheck.setGeometry(QtCore.QRect(110, 360, 100, 31))
-        self.createFunctionsButton.setGeometry(QtCore.QRect(230, 360, 181, 31))
-
         self.bg.raise_()
         self.switchModeLabel.raise_()
         self.vSepLine.raise_()
@@ -787,26 +1204,26 @@ class SkeletonWindow(QMainWindow):
         self.pdbDatabaseLabel.raise_()
         self.modeInfoHLine.raise_()
         self.infoDatabaseHLine.raise_()
-        # self.cColorLabel.raise_()
-        # self.oColorLabel.raise_()
-        # self.nColorLabel.raise_()
-        # self.sColorLabel.raise_()
-        # self.pColorLabel.raise_()
+        self.cColorLabel.raise_()
+        self.oColorLabel.raise_()
+        self.nColorLabel.raise_()
+        self.sColorLabel.raise_()
+        self.pColorLabel.raise_()
         self.backboneColorLabel.raise_()
         self.sidechainColorLabel.raise_()
         self.otherColorLabel.raise_()
-        # self.oColorBox.raise_()
-        # self.nColorBox.raise_()
-        # self.pColorBox.raise_()
+        self.oColorBox.raise_()
+        self.nColorBox.raise_()
+        self.pColorBox.raise_()
         self.otherColorBox.raise_()
-        # self.sColorBox.raise_()
-        # self.cColorBox.raise_()
+        self.sColorBox.raise_()
+        self.cColorBox.raise_()
         self.backboneColorBox.raise_()
         self.sidechainColorBox.raise_()
         self.aScaleLabel.raise_()
-        # self.showAtomsCheck.raise_()
+        self.showAtomsCheck.raise_()
         self.otherMoleculeCheck.raise_()
-        #self.meshCheck.raise_()
+        self.meshCheck.raise_()
         self.aScaleSpinBox.raise_()
         self.showBackboneCheck.raise_()
         self.showSidechainCheck.raise_()
@@ -842,6 +1259,12 @@ class SkeletonWindow(QMainWindow):
         self.selectMinecraftSaveButton.clicked.connect(self.handle_select_minecraft_button)
         self.createFunctionsButton.clicked.connect(self.handle_make_function_button)
 
+        #self.oColorBox.focusOut.connect(self.check_input)
+        self.oColorBox.focusOut.connect(lambda: self.check_input(self.oColorBox, decorative_blocks))
+        self.cColorBox.focusOut.connect(lambda: self.check_input(self.cColorBox, decorative_blocks))
+        self.sColorBox.focusOut.connect(lambda: self.check_input(self.sColorBox, decorative_blocks))
+        self.pColorBox.focusOut.connect(lambda: self.check_input(self.pColorBox, decorative_blocks))
+        self.nColorBox.focusOut.connect(lambda: self.check_input(self.nColorBox, decorative_blocks))
         self.otherColorBox.focusOut.connect(lambda: self.check_input(self.otherColorBox, decorative_blocks))
         self.backboneColorBox.focusOut.connect(lambda: self.check_input(self.backboneColorBox, decorative_blocks))
         self.sidechainColorBox.focusOut.connect(lambda: self.check_input(self.sidechainColorBox, decorative_blocks))
@@ -849,7 +1272,6 @@ class SkeletonWindow(QMainWindow):
         self.showBackboneCheck.stateChanged.connect(self.on_showBackboneCheck_changed)
         self.showSidechainCheck.stateChanged.connect(self.on_showSidechainCheck_changed)
         self.colorByBackboneCheck.stateChanged.connect(self.on_colorByBackboneCheck_changed)
-        self.otherMoleculeCheck.stateChanged.connect(self.on_otherMoleculeCheck_changed)
 
     def on_showBackboneCheck_changed(self, state):
         if not self.colorByBackboneCheck.isChecked():
@@ -864,11 +1286,6 @@ class SkeletonWindow(QMainWindow):
     def on_colorByBackboneCheck_changed(self, state):
         self.backboneColorBox.setEnabled(state == 0)
         self.sidechainColorBox.setEnabled(state == 0)
-
-    def on_otherMoleculeCheck_changed(self, state):
-        self.otherColorBox.setEnabled(state != 0)
-        self.aScaleSpinBox.setEnabled(state != 0)
-
 
     def check_input(self, combobox, valid_options):
         text = combobox.currentText()
@@ -908,24 +1325,28 @@ class SkeletonWindow(QMainWindow):
 
     def handle_make_function_button(self):
         # Create a dictionary to store the user options
-        config_data = {'atoms': {}}
+        config_data = {}
+        config_data['atoms'] = {}
 
         # Add the current text of each combobox to the dictionary
-        config_data['atoms']['O'] = 'red_wool'
-        config_data['atoms']['N'] = 'blue_wool'
-        config_data['atoms']['P'] = 'lime_wool'
-        config_data['atoms']['S'] = 'yellow_wool'
-        config_data['atoms']['C'] = 'black_wool'
+        config_data['atoms']['O'] = self.oColorBox.currentText()
+        config_data['atoms']['N'] = self.nColorBox.currentText()
+        config_data['atoms']['P'] = self.pColorBox.currentText()
+        config_data['atoms']['S'] = self.sColorBox.currentText()
+        config_data['atoms']['C'] = self.cColorBox.currentText()
         config_data['atoms']['FE'] = 'iron_block'
         config_data['atoms']['other_atom'] = self.otherColorBox.currentText()
         config_data['atoms']['backbone_atom'] = self.backboneColorBox.currentText()
         config_data['atoms']['sidechain_atom'] = self.sidechainColorBox.currentText()
+
         config_data['backbone_size'] = self.backboneScaleSpinBox.value()
         config_data['atom_scale'] = self.aScaleSpinBox.value()
         config_data['scale'] = self.pScaleSpinBox.value()
 
         # Add the checked state of each checkbox to the dictionary
+        config_data['show_atoms'] = self.showAtomsCheck.isChecked()
         config_data['show_hetatm'] = self.otherMoleculeCheck.isChecked()
+        config_data['mesh'] = self.meshCheck.isChecked()
         config_data['backbone'] = self.showBackboneCheck.isChecked()
         config_data['sidechain'] = self.showSidechainCheck.isChecked()
         config_data['by_chain'] = self.colorByBackboneCheck.isChecked()
@@ -987,13 +1408,12 @@ class SkeletonWindow(QMainWindow):
             print(hetatom_df.head())
             print(hetatm_bonds.head())
             try:
-                print("fix")
-                skeleton.run_mode(config_data, pdb_name, pdb_file, rounded, mc_dir, atom_df, hetatom_df, hetatm_bonds)
+                custom.run_mode(config_data, pdb_name, pdb_file, rounded, mc_dir, atom_df, hetatom_df, hetatm_bonds)
             except Exception as e:
                 print(f"Error: {e}")
 
             mcfiles = mcf.find_mcfunctions(mc_dir, pdb_name.lower())
-            #print(mcfiles)
+            print(mcfiles)
 
             if config_data["simple"]:
                 mcf.create_simple_function(pdb_name, mc_dir)
@@ -1022,13 +1442,13 @@ class SkeletonWindow(QMainWindow):
 
     def handle_custom_mode(self):
         print("Custom mode button clicked")
-        self.Custom = customWindow.CustomWindow()
-        self.Custom.show()
-        self.hide()
 
     def handle_skeleton_mode(self):
         print("Skeleton mode button clicked")
-
+        self.Skeleton = skeletonWindow.SkeletonWindow()
+        self.Skeleton.show()
+        #Turn off main window
+        self.hide()
 
     def handle_xray_mode(self):
         print("X-Ray mode button clicked")
@@ -1055,48 +1475,51 @@ class SkeletonWindow(QMainWindow):
         self.hide()
 
 
-    def retranslateUi(self, SkeletonWindow):
+    def retranslateUi(self, CustomWindow):
         _translate = QtCore.QCoreApplication.translate
-        #SkeletonWindow.setWindowTitle(_translate("SkeletonWindow", "MainWindow"))
-        self.switchModeLabel.setText(_translate("SkeletonWindow", "Switch Mode"))
-        self.CustomMode.setText(_translate("SkeletonWindow", "Custom"))
-        self.SkeletonMode.setText(_translate("SkeletonWindow", "Skeleton"))
-        self.XRayMode.setText(_translate("SkeletonWindow", "X-Ray"))
-        self.SpaceFillingMode.setText(_translate("SkeletonWindow", "Space Filling"))
-        self.AminoAcidMode.setText(_translate("SkeletonWindow", "Amino Acids"))
-        self.RibbonMode.setText(_translate("SkeletonWindow", "Ribbon"))
-        self.github.setText(_translate("SkeletonWindow", "Github"))
-        self.help.setText(_translate("SkeletonWindow", "Help"))
-        self.rcsbButton.setText(_translate("SkeletonWindow", "RCSB.org"))
-        self.mc2pdbLabel.setText(_translate("SkeletonWindow", "PDB2MC"))
-        self.pdbDatabaseLabel.setText(_translate("SkeletonWindow", "PDB Database"))
-        # self.cColorLabel.setText(_translate("SkeletonWindow", "Select C atom color:"))
-        # self.oColorLabel.setText(_translate("SkeletonWindow", "Select O atom color:"))
-        # self.nColorLabel.setText(_translate("SkeletonWindow", "Select N atom color:"))
-        # self.sColorLabel.setText(_translate("SkeletonWindow", "Select S atom color:"))
-        # self.pColorLabel.setText(_translate("SkeletonWindow", "Select P atom color:"))
-        self.backboneColorLabel.setText(_translate("SkeletonWindow", "Select backbone color:"))
-        self.sidechainColorLabel.setText(_translate("SkeletonWindow", "Select sidechain color:"))
-        self.otherColorLabel.setText(_translate("SkeletonWindow", "Select other color:"))
-        self.aScaleLabel.setText(_translate("SkeletonWindow", "Atom scale:"))
-        # self.showAtomsCheck.setText(_translate("SkeletonWindow", "Show model atoms"))
-        self.otherMoleculeCheck.setText(_translate("SkeletonWindow", "Show other molecules"))
-        #self.meshCheck.setText(_translate("SkeletonWindow", "Use \"mesh-style\" atoms"))
-        self.showBackboneCheck.setText(_translate("SkeletonWindow", "Show backbone"))
-        self.showSidechainCheck.setText(_translate("SkeletonWindow", "Show sidechain"))
-        self.colorByBackboneCheck.setText(_translate("SkeletonWindow", "Color backbone by chain"))
-        self.pScaleLabel.setText(_translate("SkeletonWindow", "Protein scale:"))
-        self.bScaleLabel.setText(_translate("SkeletonWindow", "Backbone scale:"))
-        self.selectIncludedPDBButton.setText(_translate("SkeletonWindow", "Select Included PDB File"))
-        self.selectMinecraftSaveButton.setText(_translate("SkeletonWindow", "Select Minecraft Save"))
-        self.simpleOutputCheck.setText(_translate("SkeletonWindow", "Simple output"))
-        self.selectPDBFileButton.setText(_translate("SkeletonWindow", "Select PDB File"))
-        self.createFunctionsButton.setText(_translate("SkeletonWindow", "Create Minecraft Functions"))
-        self.orText.setText(_translate("SkeletonWindow", "or"))
-        self.andText.setText(_translate("SkeletonWindow", "and"))
+        #CustomWindow.setWindowTitle(_translate("CustomWindow", "MainWindow"))
+        self.switchModeLabel.setText(_translate("CustomWindow", "Switch Mode"))
+        self.CustomMode.setText(_translate("CustomWindow", "Custom"))
+        self.SkeletonMode.setText(_translate("CustomWindow", "Skeleton"))
+        self.XRayMode.setText(_translate("CustomWindow", "X-Ray"))
+        self.SpaceFillingMode.setText(_translate("CustomWindow", "Space Filling"))
+        self.AminoAcidMode.setText(_translate("CustomWindow", "Amino Acids"))
+        self.RibbonMode.setText(_translate("CustomWindow", "Ribbon"))
+        self.github.setText(_translate("CustomWindow", "Github"))
+        self.help.setText(_translate("CustomWindow", "Help"))
+        self.rcsbButton.setText(_translate("CustomWindow", "RCSB.org"))
+        self.mc2pdbLabel.setText(_translate("CustomWindow", "PDB2MC"))
+        self.pdbDatabaseLabel.setText(_translate("CustomWindow", "PDB Database"))
+        self.cColorLabel.setText(_translate("CustomWindow", "Select C atom color:"))
+        self.oColorLabel.setText(_translate("CustomWindow", "Select O atom color:"))
+        self.nColorLabel.setText(_translate("CustomWindow", "Select N atom color:"))
+        self.sColorLabel.setText(_translate("CustomWindow", "Select S atom color:"))
+        self.pColorLabel.setText(_translate("CustomWindow", "Select P atom color:"))
+        self.backboneColorLabel.setText(_translate("CustomWindow", "Select backbone color:"))
+        self.sidechainColorLabel.setText(_translate("CustomWindow", "Select sidechain color:"))
+        self.otherColorLabel.setText(_translate("CustomWindow", "Select other color:"))
+        self.aScaleLabel.setText(_translate("CustomWindow", "Atom scale:"))
+        self.showAtomsCheck.setText(_translate("CustomWindow", "Show model atoms"))
+        self.otherMoleculeCheck.setText(_translate("CustomWindow", "Show other molecules"))
+        self.meshCheck.setText(_translate("CustomWindow", "Use \"mesh-style\" atoms"))
+        self.showBackboneCheck.setText(_translate("CustomWindow", "Show backbone"))
+        self.showSidechainCheck.setText(_translate("CustomWindow", "Show sidechain"))
+        self.colorByBackboneCheck.setText(_translate("CustomWindow", "Color backbone by chain"))
+        self.pScaleLabel.setText(_translate("CustomWindow", "Protein scale:"))
+        self.bScaleLabel.setText(_translate("CustomWindow", "Backbone scale:"))
+        self.selectIncludedPDBButton.setText(_translate("CustomWindow", "Select Included PDB File"))
+        self.selectMinecraftSaveButton.setText(_translate("CustomWindow", "Select Minecraft Save"))
+        self.simpleOutputCheck.setText(_translate("CustomWindow", "Simple output"))
+        self.selectPDBFileButton.setText(_translate("CustomWindow", "Select PDB File"))
+        self.createFunctionsButton.setText(_translate("CustomWindow", "Create Minecraft Functions"))
+        self.orText.setText(_translate("CustomWindow", "or"))
+        self.andText.setText(_translate("CustomWindow", "and"))
 
 if __name__ == "__main__":
     app = QApplication([])
-    main_window = SkeletonWindow()
+    main_window = CustomWindow()
     main_window.show()
-    app.exec()
+    try:
+        app.exec()
+    except KeyboardInterrupt:
+        pass
